81 errors

+++ src/main/java/slimeknights/tconstruct/library/client/data/AbstractFluidTooltipProvider.java
@@ -103,7 +103,7 @@ public abstract class AbstractFluidTooltipProvider extends GenericDataProvider {
     private final ImmutableList.Builder<FluidUnit> units = ImmutableList.builder();
 
     /** Adds a unit with a full translation key */
-    public FluidUnitListBuilder addUnitRaw(String key, int amount) {
+    public FluidUnitListBuilder addUnitRaw(String key, long amount) {
       units.add(new FluidUnit(key, amount));
       return this;
     }
@@ -114,7 +114,7 @@ public abstract class AbstractFluidTooltipProvider extends GenericDataProvider {
     }
 
     /** Adds a unit local to the current mod */
-    public FluidUnitListBuilder addUnit(String key, int amount) {
+    public FluidUnitListBuilder addUnit(String key, long amount) {
       return addUnitRaw(Util.makeTranslationKey("gui", id("fluid." + key)), amount);
     }
